Implement reusable get company/user id from url
