+++ frontend/src/app/announcements/announcements.component.ts
@@ -3,7 +3,11 @@ import axios from 'axios';
 import { AnnouncementDto } from '../interfaces';
 import { from } from 'rxjs';
 import { Router } from '@angular/router';
-import { getFullUser } from '../http-requests';
+import {
+  getCompanyIdFromUrl,
+  getFullUser,
+  getUserIdFromUrl,
+} from '../utility-functions';
 
 @Component({
   selector: 'app-announcements',
@@ -28,20 +32,10 @@ export class AnnouncementsComponent implements OnInit {
     if (localStorage.getItem('isLoggedIn') !== 'true') {
       this.router.navigate(['/']);
     }
-    this.getIdsFromUrl();
+    this.userId = getUserIdFromUrl();
+    this.companyId = getCompanyIdFromUrl();
     this.user = await getFullUser(this.userId);
     this.getAnnouncements();
-  }
-
-  getIdsFromUrl() {
-    const url = location.href;
-    const userMatch = url.match(/\/user\/(\d+)\//);
-    const companyMatch = url.match(/\/company\/(\d+)\//);
-
-    if (userMatch && companyMatch) {
-      this.userId = userMatch[1];
-      this.companyId = companyMatch[1];
-    }
   }
 
   async getAnnouncements() {
